# server/openapi.yaml
openapi: 3.0.3

info:
  title: pi-flashcards-server
  version: 0.1.0
  description: |
    A tiny language-learning flashcards API for the Sandy Bank homelab
    fleet. This version manages the vocabulary deck: list Terms, add a
    Term, edit a Term's notes, delete a Term. Terms are persisted in
    SQLite and survive a restart.

    A Term's three text fields (`foreign_lang`, `foreign_text`,
    `pivot_text`) are its identity: the `id` is a deterministic UUIDv5
    derived from them, so re-posting the same Term returns the existing
    row rather than creating a duplicate. Only `notes` is editable.

    Creating a Term also creates its two Cards — one prompting with the
    foreign text (`prompt_side: foreign`, recognition), one with the pivot
    text (`prompt_side: pivot`, production) — each on a Leitner schedule.
    `GET /cards` lists them; `POST /cards/{id}/reviews` grades an attempt
    and reschedules the Card, appending to an append-only review log.
    Deleting a Term removes its Cards and their reviews.

    Response-body convention:
      - endpoints about one Term return a `Term`
      - endpoints about many Terms return a bare `[Term, ...]`
      - every error returns `{ "error": "<message>" }`
      - success is always `200`
      - `GET /healthz` is the exception: `text/plain` `"ok"`

# No authentication: the service binds to loopback and is only reachable
# via the shared nginx proxy on the LAN (see pi-flashcards-server.service).
security: []

# Relative on purpose: resolves against wherever the spec was fetched
# from. The shared Swagger UI loads it as /flashcards/openapi.yaml, and
# the API is mounted one level down at /flashcards/api/ (nginx strips
# that prefix before proxying), so "Try it out" hits the right origin
# whether the box is reached as pi-0.local or a raw IP.
servers:
  - url: /flashcards/api
    description: Through the shared nginx reverse proxy, which strips the /flashcards/api/ prefix
  - url: http://localhost:8081
    description: Direct to the process, as bound by BIND_ADDR (127.0.0.1:8081 under systemd)

paths:
  /healthz:
    get:
      operationId: healthz
      summary: Liveness check
      responses:
        "200":
          description: Always "ok" while the process is up
          content:
            text/plain:
              schema:
                type: string
                example: ok

  /terms:
    get:
      operationId: listTerms
      summary: List every Term
      responses:
        "200":
          description: All Terms, oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Term"
    post:
      operationId: createTerm
      summary: Add a Term
      description: |
        Idempotent in the id: posting a Term whose three texts match an
        existing Term returns that Term unchanged, with no new row.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTerm"
      responses:
        "200":
          description: The Term as stored, with its derived id and created_at
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Term"
        "400":
          description: A text field was empty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /terms/{id}:
    patch:
      operationId: patchTermNotes
      summary: Edit a Term's notes
      description: The three text fields are immutable identity; only `notes` can change.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotesPatch"
      responses:
        "200":
          description: The updated Term
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Term"
        "400":
          description: Malformed request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No Term with that id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteTerm
      summary: Delete a Term
      description: Cascades — the Term's two Cards and all their reviews go with it.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The id that was deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deleted"
        "404":
          description: No Term with that id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /cards:
    get:
      operationId: listCards
      summary: List practice Cards
      description: |
        Every Card by default. `due_before` (ISO-8601) keeps only Cards
        due at or before that instant; `limit` caps the count. The
        practice screen calls it with `due_before=<now>&limit=20` to pull
        its queue. Results are ordered by `due_at`, oldest first.
      parameters:
        - name: due_before
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Keep only Cards with `due_at <= due_before`
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          description: Cap on the number of Cards returned
      responses:
        "200":
          description: The matching Cards, oldest-due first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PracticeCard"
        "400":
          description: "`due_before` is not a valid ISO-8601 timestamp"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /cards/{id}/reviews:
    post:
      operationId: createReview
      summary: Grade one attempt at a Card
      description: |
        In one transaction: append a `review` row, run the `Scheduler` to
        get the Card's next `schedule_state` and `due_at`, write them
        back, and return the updated Card. A `pass` promotes the Card a
        Leitner box and pushes `due_at` out; a `fail` resets it to box 1.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewReview"
      responses:
        "200":
          description: The Card after rescheduling
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeCard"
        "400":
          description: Malformed body or a `rating` outside `pass`/`fail`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No Card with that id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Term:
      type: object
      required: [id, foreign_lang, foreign_text, pivot_text, notes, created_at]
      properties:
        id:
          type: string
          format: uuid
          description: Deterministic UUIDv5 over the three text fields
          example: 9f1c1c8a-1b2d-5e3f-8a4b-6c7d8e9f0a1b
        foreign_lang:
          type: string
          description: ISO 639-1 code of the language being learned
          example: es
        foreign_text:
          type: string
          example: perro
        pivot_text:
          type: string
          example: dog
        notes:
          type: string
          nullable: true
          description: The only editable field
          example: el perro (m)
        created_at:
          type: string
          format: date-time
    NewTerm:
      type: object
      required: [foreign_lang, foreign_text, pivot_text]
      properties:
        foreign_lang:
          type: string
          minLength: 1
          example: es
        foreign_text:
          type: string
          minLength: 1
          example: perro
        pivot_text:
          type: string
          minLength: 1
          example: dog
        notes:
          type: string
          nullable: true
          example: el perro (m)
    NotesPatch:
      type: object
      required: [notes]
      properties:
        notes:
          type: string
          nullable: true
          description: New notes value; null clears them
    Deleted:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: string
          format: uuid
          description: The id of the Term that was removed
    PracticeCard:
      type: object
      description: A Card with its Term's text resolved for display during practice.
      required:
        [id, term_id, prompt_side, prompt, answer, notes, due_at, box]
      properties:
        id:
          type: string
          format: uuid
          description: Deterministic UUIDv5 over `term_id ␟ prompt_side`
        term_id:
          type: string
          format: uuid
        prompt_side:
          type: string
          enum: [foreign, pivot]
          description: |
            `foreign` shows the foreign text and asks for the pivot text
            (recognition); `pivot` is the reverse (production).
        prompt:
          type: string
          description: The text shown — `foreign_text` or `pivot_text` per `prompt_side`
          example: gato
        answer:
          type: string
          description: The text to recall — the other side
          example: cat
        notes:
          type: string
          nullable: true
          description: The parent Term's notes
        due_at:
          type: string
          format: date-time
          description: When the Card next falls due
        box:
          type: integer
          minimum: 1
          description: Current Leitner box, read from the Card's schedule state
          example: 1
    NewReview:
      type: object
      required: [rating]
      properties:
        rating:
          type: string
          enum: [pass, fail]
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: foreign_text must not be empty